[PCC-14] Feat: Add create Pizza screen

+++ app/src/main/java/com/github/komidawi/pizzacostcalculator/model/PizzaModel.kt
@@ -1,9 +1,7 @@
 package com.github.komidawi.pizzacostcalculator.model
 
-import java.math.BigDecimal
-
 data class PizzaModel(
     val name: String,
-    val size: BigDecimal,
-    val price: BigDecimal
+    val size: Float,
+    val price: Float
 )
